Fix compilation errors and remove Tauri references
- audio.rs: Fix thread_local!/RefCell usage (.with() instead of .lock())
- audio.rs: Fix field access to use self.inner
- gpu.rs: Fix DeviceNotFound type mismatch (u32 → usize)
- video_filters.rs: Add missing wgpu::util::DeviceExt import
- streaming.rs: Fix temporary string lifetime issues in FFmpeg args
- Cargo.toml: Add rayon as optional feature/dependency
- Rename "Tauri Commands" sections to "Public API" across codebase

+++ slain-core/src/gpu.rs
@@ -556,7 +556,7 @@ impl NvapiLoader {
         
         let status = unsafe { enum_fn(handles.as_mut_ptr(), &mut count) };
         if status != NVAPI_OK || device_index >= count {
-            return Err(GpuError::DeviceNotFound(device_index));
+            return Err(GpuError::DeviceNotFound(device_index as usize));
         }
         
         let handle = handles[device_index as usize];
